changed footer component to carry useeffect and fetch request logic from home,wellbeing and technical pages. this is to reduce repeated code.
