+++ src/Components/Home/Home.js
@@ -1,31 +1,10 @@
 import { useNavigate } from "react-router-dom";
-import { useEffect, useState } from "react";
 import "./Home.css";
 import Card from "../Card/Card";
 import Footer from "../Footer/Footer";
 import Logo from "../Logo/Logo";
 
 const Home = (props) => {
-  const [quote, setQuote] = useState("");
-  const [author, setAuthor] = useState("");
-
-  useEffect(() => {
-    async function fetchData() {
-      async function randomQuote() {
-        const id = Math.floor(Math.random() * 1000).toString();
-        const response = await fetch(`http://localhost:5432/quotes/${id}`);
-        console.log(response);
-        const data = await response.json();
-        console.log(data);
-        setQuote(data.payload[0].text);
-        setAuthor(data.payload[0].author);
-      }
-      randomQuote();
-    }
-    fetchData();
-    console.log("useEffect");
-  }, []);
-
   const navigate = useNavigate();
   return (
     <div className="Home">
@@ -51,9 +30,6 @@ const Home = (props) => {
       </div>
       <div className="footer-class">
         <Footer />
-        <p className="quote-paragraph">
-          {quote} {author}
-        </p>
       </div>
     </div>
   );
