+++ src/Components/Wellbeing/Wellbeing.js
@@ -1,31 +1,9 @@
-import { useEffect, useState } from "react";
 import Section from "../Section/Section";
 import "../Wellbeing/Wellbeing.css";
 import Footer from "../Footer/Footer";
 import Logo from "../Logo/Logo";
 
 const Wellbeing = () => {
-  const [quote, setQuote] = useState("");
-  const [author, setAuthor] = useState("");
-
-  useEffect(() => {
-    async function fetchData() {
-      async function randomQuote() {
-        const id = Math.floor(Math.random() * 1000);
-        const response = await fetch(
-          `http://localhost:5432/quotes/${id.toString()}`
-        );
-        console.log(response);
-        const data = await response.json();
-        console.log(data);
-        setQuote(data.payload[0].text);
-        setAuthor(data.payload[0].author);
-      }
-      randomQuote();
-    }
-    fetchData();
-    console.log("useEffect");
-  }, []);
   return (
     <div className="wellbeing-div">
       <div className="title-section">
@@ -81,9 +59,6 @@ const Wellbeing = () => {
       </div>
       <div className="footer-class">
         <Footer />
-        <p className="quote-paragraph">
-          {quote} {author}
-        </p>
       </div>
     </div>
   );
